refactor graphics
