+++ cocos2d/core/renderer/assemblers/graphics/earcut.js
