+++ cocos2d/core/renderer/assemblers/index.js
@@ -1,1 +1,2 @@
 require('./sprite-assembler');
+require('./graphics/graphics-assembler');
